refactor(day08-oop): 重构 Lambda 表达式示例代码
- 移动 Demo06、Student 和 SwimAble 类到 com.itheima._05Lambda 包
- 新增 Demo07 类，演示使用 Lambda 表达式进行数组排序- 更新现有类以适应新包结构
